refactor OrderCreation tests
